requirements.txt
pandas==2.2.3
scipy==1.15.2
fancyimpute==0.7.0
tensorflow-gpu==2.11.0
numpy==1.26.4